+++ README.md
@@ -1,2 +1,13 @@
-# css-frameworks-ca
-OOOOGA BOOOOGA BOOOOOGA
+# js2-ca
+
+## requirements:
+
+1. X User with @noroff.no or @stud.noroff.no email can register profile
+2. X Registered user can login
+3. User can view a post content feed
+4. User can filter the post content feed
+5. User can search the post content feed
+6. User can view a post content item by ID
+7. User can create a post content item
+8. User can update a post content item
+9. User can delete a post content item
