+++ feed/index.html
@@ -11,6 +11,7 @@
 	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
 		integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
 		crossorigin="anonymous"></script>
+	<script src="../js/feed.mjs" type="module"></script>
 	<title>feed | meetMarket</title>
 </head>
 
@@ -40,7 +41,7 @@
 							<a class="btn btn-outline-primary my-auto mx-2 nav-link" href="../profile/">My
 								Profile</a>
 						</li>
-						<li class="nav-item visually-hidden">
+						<li class="nav-item visually-hidden" id="logOutBtn">
 							<btn class="btn btn-outline-primary my-auto mx-2 nav-link">Log out</btn>
 						</li>
 					</ul>
