+++ index.html
@@ -40,6 +40,9 @@
 							<a class="btn btn-outline-primary my-auto mx-2 nav-link" href="./profile/">My
 								Profile</a>
 						</li>
+						<li class="nav-item visually-hidden" id="logOutBtn">
+							<btn class="btn btn-outline-primary my-auto mx-2 nav-link">Log out</btn>
+						</li>
 					</ul>
 				</div>
 			</div>
