+++ js/feed.mjs
@@ -1,0 +1,3 @@
+import { logoutBtn } from "./utils/logoutBtn.js";
+
+logoutBtn();
